+++ changelog.md
@@ -1,3 +1,9 @@
+## Changes in 1.2.0 : ##
+
+- Supports GeoSpatial index.
+- Validates resource id.
+- Adds new header "index transformation progress" to ResourceResponse.
+
 ## Changes in 1.1.0 : ##
 
 - Implements V2 indexing policy